refact : ProfilePage를 클라이언트/서버 컴포넌트로 분리 및  GroupPage 'use client' 제거
