+++ src/app/group/page.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import React from 'react';
 import classNames from 'classnames/bind';
 import Image from 'next/image';
